Refactor of user profile
