--- src/main/java/com/szymkowski/rpg/security/UserRepository.java
@@ -1,8 +1,0 @@
-package com.szymkowski.rpg.security;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface UserRepository extends JpaRepository<User, Long> {
-
-    User findDistinctByEmail(String email);
-}
