+++ src/main/java/com/szymkowski/rpg/user/UserRepository.java
@@ -1,0 +1,9 @@
+package com.szymkowski.rpg.user;
+
+import com.szymkowski.rpg.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findDistinctByEmail(String email);
+}
