+++ src/main/java/com/szymkowski/rpg/user/UserResource.java
@@ -1,5 +1,6 @@
-package com.szymkowski.rpg.security;
+package com.szymkowski.rpg.user;
 
+import com.szymkowski.rpg.user.User;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
